requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.1
pydantic==2.8.2
SQLAlchemy==2.0.30
alembic==1.13.2
Jinja2==3.1.4
APScheduler==3.10.4
pandas==2.2.2
email-validator==2.2.0
google-generativeai==0.7.2
httpx==0.27.0
aiosmtplib==3.0.1
